현품박스 오류 및 로직 변경

+++ SmartMES_Giroei/P1C/P1C01_PROD_ORDER.cs
@@ -152,6 +152,15 @@ namespace SmartMES_Giroei
             var data = sql;
             Logger.ApiLog(G.UserID, lblTitle.Text, ActionType.조회, data);
 
+            sql = "update SAL_order_sub set prod_status = 2 where order_id = '" + sSujuNo + "' and order_seq = " + sSujuSeq;
+            m.dbCUD(sql, ref msg);
+
+            if (msg != "OK")
+            {
+                lblMsg.Text = msg;
+                return;
+            }
+
             sql = "delete from PRD_prod_order where job_no = '" + sLotNo + "'";
             m.dbCUD(sql, ref msg);
